Creature Base Compiles
Went through the largest class EntityCreatureBase, most things are intact but pathing and AI is going to take some re-learning and I also need to look into the EntitySize class and EntityType. Perhaps CreatureInfo can manage EntityType.

+++ src/main/java/com/lycanitesmobs/core/entity/ai/DirectNavigator.java
@@ -130,7 +130,7 @@ public class DirectNavigator {
 	// ==================================================
   	//                      Movement
   	// ==================================================
-	public void flightMovement(float moveStrafe, float moveForward) {
+	public void flightMovement(double moveStrafe, double moveForward) {
 		if(this.host.isInWater()) {
             this.host.moveSwimmingWithHeading(moveStrafe, moveForward);
             this.host.move(MoverType.SELF, this.host.motionX, this.host.motionY, this.host.motionZ);
